* import-visit-2.12.2:
  databases: add in the paraDIS_tecplot reader
  databases: update for the upstream removal of PLOT2D
  visit-databases 2017-04-27 (3a3a068a)
  scripts: add a script for importing the database readers
  databases: remove old imports
  databases: move imported readers into their own directory
  Vs: add a missing newline
  utf-8: fix invalid utf-8 sequences
  databases: remove link to avtivp
  visit 2017-04-27 (3a3a068a)
  VisItLib: remove the old avt import
  scripts: add a script for importing the core AVT directory
  scripts: add `update-common.sh` for importing code from a repo

Merge-request: !34

+++ Library/VisItLib/avt/DBAtts/SIL/avtSILArray.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -452,12 +452,12 @@ avtSILArray::GetSetIndex(const std::string &name) const
     }
     if (names.size() != 0)
     {
-        int ii = 0;
+        size_t ii = 0;
         for (ii = 0; ii < names.size(); ii++)
         {
             if (names[ii] == name)
             {
-                return (iFirstSet + ii);
+                return (iFirstSet + (int)ii);
             }
         }
         return -1;
@@ -504,7 +504,7 @@ avtSILArray::IsCompatible(const avtSILArray *a) const
         return false;
     if (names.size() != a->names.size())
         return false;
-    for (int i = 0 ; i < names.size() ; i++)
+    for (size_t i = 0 ; i < names.size() ; i++)
     {
         if (names[i] != a->names[i])
             return false;
